// src/api/types.ts
export type Character = {
  id: number
  name: string
  status: 'Alive' | 'Dead' | 'unknown'
  species: string
  type: string
  gender: string
  origin: { name: string; url: string }
  location: { name: string; url: string }
  image: string
  episode: string[]
  url: string
  created: string
}

export type Episode = {
  id: number
  name: string
  air_date: string
  episode: string
  characters: string[]
  url: string
  created: string
}

export type PageInfo = {
  count: number
  pages: number
  next: string | null
  prev: string | null
}

export type CharactersResponse = {
  info: PageInfo
  results: Character[]
}